Create additional incomings for localhost
In the old days, multiserver servers bound to all interfaces. Now they only
bind to hosts that are explicitly configured in either config.host or
config.connections.incoming.**.host, or the host indirectly selected via a scope.

The old behaviour of binding to all interfaces was convenient, because
sbot would gossip on the LAN/Internet and would also allow the local cli client to
connect without any configuration.

This resores this convenient behaviour by creating additional local incomings
for net and ws as neccessary.